feat: add local file upload
Feature also inclides additional dependencies to (`react-pdf` and `raw-loader`) to render the locally selected file on the web UI.

Note, this is purely a frontend implementation stage at the moment, where the selected file for upload is only stored in memory temporarily (frontend state). Will implement a backend API later to receive and consume the file and parse this in a Python script via FastAPI.

Also, refactored some configurations, including shorthand paths.