chore: update npm run script to activate python environment
Previously when running the full stack server via NPM, the python packages were installed globally, outside of the project virtual environment. Sometimes we forget to activate the python virtual environment...

This change should help ensure that we run the server and install python packages within the virtual environment whenever we run the servers.